better error on no size blob restore

+++ sync.py
@@ -23,6 +23,8 @@ if action == "backup":
 elif action.startswith("restore"):
     # head request to a blob URL returns its size
     size = check_output(f"curl -s --head {remoteurl}" + " | awk '$1 == \"Content-Length:\" {print $2}' | tr -d '\r' | numfmt --to=iec", shell=True).decode("utf8").strip()
+    if not size:
+        exit(f"Remote blob '{remote}' does not exist...")
     print(f"Restoring '{remote}' ({size} tar.lz4) to '{local}'")
     if os.path.exists("lost+found"): # handle new filesystems
         os.rmdir("lost+found")
